parent 13bf24f052a54b04e1e658c639e4072ceb8c3bc1 author mickael-caro-sonarsource <mickael.caro@sonarsource.com> 1572620791 +0100 committer Mickaël <mickael.caro@sonarsource.com> 1572870174 +0100
Removed the DEV version in ITs, because of breaking changes in the API that will be supported in the scanner later.

Bumped Tests projects to .NET Framework 4.7.2, and updated packages according, especially MIcrosoft.Build, to version 16.3.0 to enable support of MSBuild 16.0.

Removed vsTestVersion previously added.

Added logging of component and issues to ITs (#802)

Deleted .cix.yml and cix.sh

Removed additional properties for nuget pack